+++ src/python_framework_v02/troute/network/musking/mc_reach.pxd
@@ -9,8 +9,8 @@ cdef extern from "mc_reach_structs.c":
 cdef extern from "mc_reach_structs.h":
   ctypedef struct _MC_Segment:
     long id;
-    float dt, dx, bw, tw, twcc, n, ncc, cs, s0;
-    float qdp, velp, depthp;
+    double dt, dx, bw, tw, twcc, n, ncc, cs, s0;
+    double qdp, velp, depthp;
   ctypedef struct _MC_Reach:
     pass
   ctypedef struct _Reach:
@@ -21,8 +21,8 @@ cdef class MC_Segment(Segment):
     A muskingcung segment
   """
   #TODO document these attributes
-  cdef readonly float dt, dx, bw, tw, twcc, n, ncc, cs, s0
-  cdef readonly float qdp, velp, depthp
+  cdef readonly double dt, dx, bw, tw, twcc, n, ncc, cs, s0
+  cdef readonly double qdp, velp, depthp
 
 cdef class MC_Reach(Reach):
   """
