+++ src/python_framework_v02/troute/network/musking/mc_reach.pyx
@@ -7,9 +7,9 @@ cdef extern from "mc_reach_structs.c":
   void init_mc_reach(_Reach* reach, int num_segments)
   void free_mc_reach(_Reach* reach)
   void set_mc_segment(_Reach* reach, int index, long id,
-      float dt, float dx, float bw, float tw, float twcc,
-      float n, float ncc, float cs, float s0,
-      float qdp, float velp, float depthp)
+      double dt, double dx, double bw, double tw, double twcc,
+      double n, double ncc, double cs, double s0,
+      double qdp, double velp, double depthp)
   _MC_Segment get_mc_segment(_Reach* reach, int index) nogil
 
 cdef class MC_Segment(Segment):
