+++ src/python_framework_v02/troute/network/reservoirs/hybrid/hybrid_structs.h
@@ -10,10 +10,10 @@
 
 typedef struct {
   int lake_number;
-  float dam_length, area, max_depth;
-  float orifice_area, orifice_coefficient, orifice_elevation;
-  float weir_coefficient, weir_elevation, weir_length;
-  float initial_fractional_depth, water_elevation;
+  double dam_length, area, max_depth;
+  double orifice_area, orifice_coefficient, orifice_elevation;
+  double weir_coefficient, weir_elevation, weir_length;
+  double initial_fractional_depth, water_elevation;
   int reservoir_type;
   char* reservoir_parameter_file;
   char* start_date;
