+++ src/python_framework_v02/troute/network/reservoirs/rfc/rfc.pyx
@@ -9,17 +9,17 @@ Externally defined symbols
 
 cdef extern from "rfc_structs.c":
   void init_rfc_reach(_Reach* reach, int lake_number,
-                            float dam_length, float area, float max_depth,
-                            float orifice_area, float orifice_coefficient, float orifice_elevation,
-                            float weir_coefficient, float weir_elevation, float weir_length,
-                            float initial_fractional_depth, float water_elevation,
+                            double dam_length, double area, double max_depth,
+                            double orifice_area, double orifice_coefficient, double orifice_elevation,
+                            double weir_coefficient, double weir_elevation, double weir_length,
+                            double initial_fractional_depth, double water_elevation,
                             int reservoir_type, char *reservoir_parameter_file, char *start_date,
                             char *time_series_path, int forecast_lookback_hours
   )
   void free_rfc_reach(_Reach* reach)
-  void route(_Reach* reach, float routing_period, float inflow, float lateral_inflow, float* outflow,  float* water_elevation) nogil
+  void route(_Reach* reach, double routing_period, double inflow, double lateral_inflow, double* outflow,  double* water_elevation) nogil
 
-cdef void run_rfc_c(_Reach* reach, float inflow, float lateral_inflow, float routing_period, float* outflow, float* water_elevation) nogil:
+cdef void run_rfc_c(_Reach* reach, double inflow, double lateral_inflow, double routing_period, double* outflow, double* water_elevation) nogil:
     route(reach, inflow, lateral_inflow, routing_period, outflow, water_elevation)
 
 cdef class MC_RFC(Reach):
@@ -102,24 +102,24 @@ cdef class MC_RFC(Reach):
     """
     free_rfc_reach(&self._reach)
 
-  cpdef (float,float) run(self, float inflow, float lateral_inflow, float routing_period):
+  cpdef (double,double) run(self, double inflow, double lateral_inflow, double routing_period):
     """
       Run the rfc routing function
       Params:
-        inflow: float
+        inflow: double
           inflow into the reservoir
-        lateral_inflow: float
+        lateral_inflow: double
           lateral flows into the reservoir
-        routing_period: float
+        routing_period: double
           amount of time to simulatie reservoir operation for, outflow if valid until this time
       Return:
-        outflow: float
+        outflow: double
           flow rate out of the reservoir valid for routing_period seconds
         water_elevation:
           reservoir water surface elevation after routing_period seconds
     """
-    cdef float outflow = 0.0
-    cdef float water_elevation = 0.0
+    cdef double outflow = 0.0
+    cdef double water_elevation = 0.0
     with nogil:
       route(&self._reach, inflow, lateral_inflow, routing_period, &outflow,  &water_elevation)
       #printf("outflow: %f\n", outflow)
